initial implementation of activity tracking

+++ sepal/organizations/lib.py
@@ -7,8 +7,8 @@ import httpx
 from fastapi import Depends, Path
 from sqlalchemy.orm import object_session
 
+import sepal.db as db
 from sepal.auth import get_current_user
-from sepal.db import Session
 from sepal.profile.lib import get_profile
 from sepal.profile.models import Profile
 from sepal.invitations.models import Invitation
@@ -31,14 +31,12 @@ class OrganizationsPermission(str, Enum):
 
 @contextmanager
 def organization_query():
-    with Session() as session:
-        yield session.query(Organization)
+    yield db.Session().query(Organization)
 
 
 @contextmanager
 def organization_user_query():
-    with Session() as session:
-        yield session.query(OrganizationUser)
+    yield db.Session().query(OrganizationUser)
 
 
 async def verify_org_id(
@@ -54,14 +52,12 @@ async def verify_org_id(
 
 
 async def is_member(org_id: int, user_id: str, role: Optional[RoleType] = None) -> bool:
-    with Session() as session:
-        q = session.query(OrganizationUser).filter_by(
-            organization_id=org_id, user_id=user_id
-        )
+    with organization_user_query() as q:
+        q = q.filter_by(organization_id=org_id, user_id=user_id)
         if role:
             q = q.filter_by(role=role)
 
-        return session.query(q.exists()).scalar()
+        return db.Session().query(q.exists()).scalar()
 
 
 async def get_organization_by_id(
@@ -96,39 +92,34 @@ async def get_user_role(org_id: int, user_id: str) -> Optional[RoleType]:
 
 async def assign_role(org_id: int, user_id: str, role: RoleType):
     """Assign a user to a role in an organization."""
-    with Session() as session:
-        org_user_exists = (
-            session.query(OrganizationUser)
-            .filter_by(organization_id=org_id, user_id=user_id, role=role)
-            .exists()
-        )
+    with organization_user_query() as q:
+        org_user_exists = q.filter_by(
+            organization_id=org_id, user_id=user_id, role=role
+        ).exists()
+        session = db.Session()
         if session.query(org_user_exists).scalar():
             return
 
         org_user = OrganizationUser(organization_id=org_id, user_id=user_id, role=role)
         session.add(org_user)
-        session.commit()
+        session.flush()
 
 
 async def remove_role(org_id: int, user_id: str, role: RoleType):
     """Remove a user from a role in an organization."""
-    with Session() as session:
-        session.query(OrganizationUser).filter_by(
-            organization_id=org_id, user_id=user_id
-        ).delete()
-        session.commit()
+    with organization_user_query() as q:
+        q.filter_by(organization_id=org_id, user_id=user_id).delete()
+        db.Session().flush()
 
 
 async def create_organization(user_id: str, data: OrganizationCreate) -> Organization:
-    with Session() as session:
-        org = Organization(**data.dict())
-        org_user = OrganizationUser(
-            organization=org, user_id=user_id, role=RoleType.Owner
-        )
-        session.add_all([org, org_user])
-        session.commit()
-        session.refresh(org)
-        return org
+    org = Organization(**data.dict())
+    org_user = OrganizationUser(organization=org, user_id=user_id, role=RoleType.Owner)
+    session = db.Session()
+    session.add_all([org, org_user])
+    session.flush()
+    session.refresh(org)
+    return org
 
 
 async def update_organization(
@@ -142,57 +133,58 @@ async def update_organization(
         for key, value in data.dict().items():
             setattr(org, key, value)
 
-        session = object_session(org)
-        session.commit()
+        session = db.Session()
+        session.flush()
         session.refresh(org)
         return org
 
 
 async def get_users(org_id: int) -> List[Tuple[Profile, RoleType]]:
-    with Session() as session:
-        return (
-            session.query(Profile, OrganizationUser.role)
-            .filter(
-                OrganizationUser.user_id == Profile.user_id,
-                OrganizationUser.organization_id == org_id,
-            )
-            .all()
+    return (
+        db.Session()
+        .query(Profile, OrganizationUser.role)
+        .filter(
+            OrganizationUser.user_id == Profile.user_id,
+            OrganizationUser.organization_id == org_id,
         )
+        .all()
+    )
 
 
 async def invite_user(org_id: int, invited_by_user_id: str, email: str):
-    with Session() as session:
-        # TODO: if the user already has a user account/profile maybe we just add
-        # them directly to the organization and send them an email telling
-        # them
-        token = secrets.token_urlsafe(16)
-        invitation = Invitation(
-            organization_id=org_id,
-            invited_by=invited_by_user_id,
-            email=email,
-            token=token,
+    # TODO: if the user already has a user account/profile maybe we just add
+    # them directly to the organization and send them an email telling
+    # them
+    token = secrets.token_urlsafe(16)
+    invitation = Invitation(
+        organization_id=org_id,
+        invited_by=invited_by_user_id,
+        email=email,
+        token=token,
+    )
+
+    session = db.Session()
+    session.add(invitation)
+    session.flush()
+
+    profile = await get_profile(invited_by_user_id)
+    org = await get_organization_by_id(org_id)
+
+    template = get_template("invitation.jinja2.html")
+    async with httpx.AsyncClient() as client:
+        resp = await client.post(
+            settings.mailgun_api_url,
+            auth=("api", settings.mailgun_api_key),
+            data={
+                "from": "Sepal Support <noreply@sepal.app>",
+                "subject": "You have been invited to join Sepal",
+                "html": template.render(
+                    url=f"{settings.app_base_url}/register?invitation={token}",
+                    invited_by=profile.email,
+                    organization_name=org.name,
+                ),
+                "to": email,
+            },
         )
-        session.add(invitation)
-        session.commit()
 
-        profile = await get_profile(invited_by_user_id)
-        org = await get_organization_by_id(org_id)
-
-        template = get_template("invitation.jinja2.html")
-        async with httpx.AsyncClient() as client:
-            resp = await client.post(
-                settings.mailgun_api_url,
-                auth=("api", settings.mailgun_api_key),
-                data={
-                    "from": "Sepal Support <noreply@sepal.app>",
-                    "subject": "You have been invited to join Sepal",
-                    "html": template.render(
-                        url=f"{settings.app_base_url}/register?invitation={token}",
-                        invited_by=profile.email,
-                        organization_name=org.name,
-                    ),
-                    "to": email,
-                },
-            )
-
-            resp.raise_for_status()
+        resp.raise_for_status()
